+++ tests/testthat/test_nafta.R
@@ -1,21 +1,3 @@
-# Copyright (C) 2019 Johannes Ranke
-# Contact: jranke@uni-bremen.de
-
-# This file is part of the R package mkin
-
-# mkin is free software: you can redistribute it and/or modify it under the
-# terms of the GNU General Public License as published by the Free Software
-# Foundation, either version 3 of the License, or (at your option) any later
-# version.
-
-# This program is distributed in the hope that it will be useful, but WITHOUT
-# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-# FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
-# details.
-
-# You should have received a copy of the GNU General Public License along with
-# this program. If not, see <http://www.gnu.org/licenses/>
-
 context("Evaluations according to 2015 NAFTA guidance")
 
 test_that("Data for more than one compound are rejected",
@@ -40,6 +22,7 @@ test_that("Test data from Appendix B are correctly evaluated", {
 
   plot_nafta <- function() plot(res)
   if(requireNamespace("vdiffr", quietly = TRUE)) {
+    skip_if(getRversion() > 4.0)
     vdiffr::expect_doppelganger("NAFTA SOP Appendix B", plot_nafta)
   }
 })
@@ -64,6 +47,7 @@ test_that("Test data from Appendix D are correctly evaluated", {
 
   plot_nafta <- function() plot(res)
   if(requireNamespace("vdiffr", quietly = TRUE)) {
+    skip_if(getRversion() > 4.0)
     vdiffr::expect_doppelganger("Plot NAFTA analysis", plot_nafta)
   }
 })